requirements.txt
PyYAML==6.0.1
openai==1.23.2
discord.py==2.3.2
python-dotenv==1.0.1
colorama==0.4.6
aiosqlite==0.20.0